bug fix for limit order

+++ server/config/limitOrder.js
@@ -13,24 +13,26 @@ module.exports.limitOrder = function(){
           list+=orders[i].dataValues.symbol + '+';
         }
         http.get('http://finance.yahoo.com/d/quotes.csv?s=' + list + '&f=sa', function(err, res){
-          var ask = res.buffer.toString().split('\n');
-          for(var i=0; i<orders.length; i++){
-            //if set price is greater than current price, process trade
-            if(Number(ask[i][1].split(',')[1]) < orders[i].dataValues.price){
-              Order.destroy({ where: {id: orders[i].id }})
-              var trade = {
-                 symbol: orders[i].dataValues.symbol,
-                 company: orders[i].dataValues.company,
-                 portfolioId: orders[i].dataValues.PortfolioId,
-                 shares: orders[i].dataValues.shares,
-                 price: orders[i].dataValues.price,
-                 marketPrice: orders[i].dataValues.price,
-                 buysell: orders[i].dataValues.buysell,
-                 executed: true,
-                 dayorder: orders[i].dataValues.dayorder
+          if(res.buffer){
+            var ask = res.buffer.toString().split('\n');
+            for(var i=0; i<orders.length; i++){
+              //if set price is greater than current price, process trade
+              if(Number(ask[i][1].split(',')[1]) < orders[i].dataValues.price){
+                Order.destroy({ where: {id: orders[i].id }})
+                var trade = {
+                   symbol: orders[i].dataValues.symbol,
+                   company: orders[i].dataValues.company,
+                   portfolioId: orders[i].dataValues.PortfolioId,
+                   shares: orders[i].dataValues.shares,
+                   price: orders[i].dataValues.price,
+                   marketPrice: orders[i].dataValues.price,
+                   buysell: orders[i].dataValues.buysell,
+                   executed: true,
+                   dayorder: orders[i].dataValues.dayorder
+                }
+                placeTrade(trade);
+                placeOrder(trade);
               }
-              placeTrade(trade);
-              placeOrder(trade);
             }
           }
         })
